Separate hanging wire into its own class. Create a common base class of wire entities.
